requirements.txt
requests
tldextract
lxml
python-dateutil
beautifulsoup4
openai
python-dotenv
rake-nltk
